--- apps/hops-api-questionnaire/main/api/questionnaire/Config.kt
@@ -1,7 +1,0 @@
-package api.questionnaire
-
-import api.questionnaire.github.GithubConfig
-
-data class Config(
-    val github: GithubConfig,
-)
