--- apps/hops-api-questionnaire/main/api/questionnaire/github/GithubConfig.kt
@@ -1,11 +1,0 @@
-package api.questionnaire.github
-
-import java.net.URL
-
-data class GithubConfig(
-    val github: Github,
-) {
-    data class Github(
-        val questionnaireUrl: URL,
-    )
-}
