+++ apps/hops-api-questionnaire/main/questionnaire/Config.kt
@@ -1,0 +1,7 @@
+package questionnaire
+
+import questionnaire.github.GithubConfig
+
+data class Config(
+    val github: GithubConfig,
+)
