+++ apps/hops-api-questionnaire/main/questionnaire/github/GithubConfig.kt
@@ -1,0 +1,11 @@
+package questionnaire.github
+
+import java.net.URL
+
+data class GithubConfig(
+    val github: Github,
+) {
+    data class Github(
+        val questionnaireUrl: URL,
+    )
+}
